Rearranged the components
Now a login page is supported
