+++ components.d.ts
@@ -1,0 +1,41 @@
+/* eslint-disable */
+// @ts-nocheck
+// Generated by unplugin-vue-components
+// Read more: https://github.com/vuejs/core/pull/3399
+// biome-ignore lint: disable
+export {}
+
+/* prettier-ignore */
+declare module 'vue' {
+  export interface GlobalComponents {
+    AlbumList: typeof import('./src/components/AlbumList.vue')['default']
+    AlbumPage: typeof import('./src/components/AlbumPage.vue')['default']
+    AlbumPreview: typeof import('./src/components/AlbumPreview.vue')['default']
+    Albums: typeof import('./src/components/Albums.vue')['default']
+    AppPage: typeof import('./src/components/AppPage.vue')['default']
+    copy: typeof import('./src/components/MainMenu copy.vue')['default']
+    HelloWorld: typeof import('./src/components/HelloWorld.vue')['default']
+    IconButton: typeof import('./src/components/IconButton.vue')['default']
+    IconCommunity: typeof import('./src/components/icons/IconCommunity.vue')['default']
+    IconDocumentation: typeof import('./src/components/icons/IconDocumentation.vue')['default']
+    IconEcosystem: typeof import('./src/components/icons/IconEcosystem.vue')['default']
+    IconSupport: typeof import('./src/components/icons/IconSupport.vue')['default']
+    IconTooling: typeof import('./src/components/icons/IconTooling.vue')['default']
+    LeftSidebar: typeof import('./src/components/LeftSidebar.vue')['default']
+    LeftSideButtons: typeof import('./src/components/LeftSideButtons.vue')['default']
+    LoginPage: typeof import('./src/components/LoginPage.vue')['default']
+    MainMenu: typeof import('./src/components/MainMenu.vue')['default']
+    MainPage: typeof import('./src/components/MainPage.vue')['default']
+    MusicPlayer: typeof import('./src/components/MusicPlayer.vue')['default']
+    PlaylistList: typeof import('./src/components/PlaylistList.vue')['default']
+    PlaylistPage: typeof import('./src/components/PlaylistPage.vue')['default']
+    PlaylistPreview: typeof import('./src/components/PlaylistPreview.vue')['default']
+    RouterLink: typeof import('vue-router')['RouterLink']
+    RouterView: typeof import('vue-router')['RouterView']
+    SearchMenu: typeof import('./src/components/SearchMenu.vue')['default']
+    SearchPage: typeof import('./src/components/SearchPage.vue')['default']
+    TheWelcome: typeof import('./src/components/TheWelcome.vue')['default']
+    TrackBar: typeof import('./src/components/TrackBar.vue')['default']
+    WelcomeItem: typeof import('./src/components/WelcomeItem.vue')['default']
+  }
+}
