+++ package.json
@@ -57,6 +57,8 @@
     "prettier": "^3.2.5",
     "tailwindcss": "^3.4.10",
     "typescript": "~5.4.0",
+    "unplugin-auto-import": "^19.1.1",
+    "unplugin-vue-components": "^28.4.1",
     "vite": "^5.3.1",
     "vite-plugin-vue-devtools": "^7.3.1",
     "vitest": "^1.6.0",
