+++ src/components/LoginPage.vue
@@ -1,0 +1,5 @@
+<script setup lang="ts"></script>
+
+<template>
+    <div></div>
+</template>
