+++ src/components/SearchMenu.vue
@@ -1,0 +1,269 @@
+<script setup lang="ts">
+import { PerfectScrollbar } from 'vue3-perfect-scrollbar';
+</script>
+
+<template>
+<div class="bg-[#362323] rounded-md overflow-hidden">
+    <form class="d-flex w-[50vh] h-[7vh] p-3 mb-3" role="search">
+        <input class="form-control me-2 input-rounded text-[#efd0d0] bg-custom" type="search" placeholder="Search" aria-label="Search">
+    </form>
+    <PerfectScrollbar class="min-h-[75vh] overflow-hidden">
+        <div class="grid grid-cols-7 gap-2">
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+        </div>
+        <div class="grid grid-cols-7 gap-2">
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+        </div>
+        <div class="grid grid-cols-7 gap-2">
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+        </div>
+        <div class="grid grid-cols-7 gap-2">
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+        </div>
+        <div class="grid grid-cols-7 gap-2">
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-full" src="../assets/images/artist.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Artist Name</div>
+                    <div class="text-white text-sm font-arial">Artist</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+            <div class="flex items-center justify-center mb-2 mt-2">
+                <div class="flex flex-col items-center">
+                    <img class="rounded-3xl" src="../assets/images/album.jpeg" width="120" height="120" style="margin-bottom: 5px;"> <!-- Adjust margin to create space between image and text -->
+                    <div class="text-white text-lg font-arial">Album Name</div>
+                    <div class="text-white text-sm font-arial">Artist Name</div>
+                </div>
+            </div>
+        </div>
+    </PerfectScrollbar>
+</div>
+
+</template>
