+++ src/main.ts
@@ -4,17 +4,17 @@ import { createApp } from 'vue'
 import { createPinia } from 'pinia'
 
 import MusicPlayer from './components/MusicPlayer.vue'
-import MainMenu from './components/MainMenu.vue'
 import LeftSidebar from './components/LeftSidebar.vue'
-import PlaylistPage from './components/PlaylistPage.vue'
+import PlaylistPage from './components/PlaylistList.vue'
 import LeftSideButtons from './components/LeftSideButtons.vue'
 import MainPage from './components/MainPage.vue'
 import SearchPage from './components/SearchPage.vue'
-import AlbumPage from './components/AlbumPage.vue'
+import AlbumPage from './components/AlbumList.vue'
 import AlbumPreview from './components/AlbumPreview.vue'
 import TrackBar from './components/TrackBar.vue'
 import PlaylistPreview from './components/PlaylistPreview.vue'
 import IconButton from './components/IconButton.vue'
+import Albums from './components/AlbumList.vue'
 
 import "bootstrap/dist/css/bootstrap.min.css"
 import "bootstrap"
@@ -25,17 +25,20 @@ import 'vue3-perfect-scrollbar/style.css';
 
 import App from './App.vue'
 import router from './router'
+import SearchMenu from './components/SearchMenu.vue'
+import PlaylistList from './components/PlaylistList.vue'
 
 const app = createApp(App)
 
 app.component('music-player', MusicPlayer)
-app.component('main-menu', MainMenu)
 app.component('left-sidebar', LeftSidebar)
-app.component('playlist-page', PlaylistPage)
+app.component('playlist-list', PlaylistList)
 app.component('left-side-buttons', LeftSideButtons)
 app.component('main-page', MainPage)
 app.component('search-page', SearchPage)
+app.component('search-menu', SearchMenu)
 app.component('album-page', AlbumPage)
+app.component('album-list', Albums)
 app.component('album-preview', AlbumPreview)
 app.component('track-bar', TrackBar)
 app.component('playlist-preview', PlaylistPreview)
